Split CastlingTest and CastlesTest
